Update tests to verify modules, and (future) link in runtime library.
The runtime library isn't linked in just yet, since there's no C API
for linking modules. I'll propose one and send a patch to LLVM.

+++ runtime/strings.go
@@ -40,7 +40,7 @@ func strcat(a, b str) str {
 	}
 
 	mem := malloc(a.size + b.size)
-	if mem == unsafe.Pointer(0) {
+	if mem == unsafe.Pointer(uintptr(0)) {
 		// TODO panic? abort?
 	}
